Exclude remove_soft_hypen and convert_latex and move to utils.py
